### 2021-06-10 #3354 @faroit
@kthyng I would be happy to edit this.

Just a disclosure: I had one [paper published](https://qmro.qmul.ac.uk/xmlui/handle/123456789/26103) with @meinardmueller (the first author of this submission) 4 years ago and we were both employed by the same institution at that time. Other than that, I don't see any part of the [JOSS COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) that would speak against me here plus I feel more than certain that I can edit this unbiased.
