### 2021-06-10 #3354 @kthyng
Thank you for stating this @faroit. I'm comfortable with you proceeding as editor in this case.
