### 2021-06-10 #3354 @faroit
👋  Hi @yiitozer @meinardmueller, thanks for your submission to JOSS! I will be serving as the editor for this (so I obviously won't review myself). We'll use this issue to identify reviewers and resolve technical issues. Thanks for the list of potential reviewers you provided. That looks good to me, I would also like to add @ magdalenafuentes if you agree that would be a good fit as well.
