### 2021-06-11 #3354 @arfon
:wave: @meinardmueller @yiitozer – I just took a cursory look at the paper and the software repository here and wondered if there was a reason for the authorship order, especially given that it looks like @yiitozer made the majority of the code contributions here?

Just to be clear, JOSS doesn't police authorship decisions, but we do flag when things look anomalous.
