### 2021-06-14 #3354 @meinardmueller
> 👋 @meinardmueller @yiitozer @ahnonay – I just took a cursory look at the paper and the software repository here and wondered if there was a reason for the authorship order, especially given that it looks like @ahnonay made the majority of the code contributions here?
>
> Just to be clear, JOSS doesn't police authorship decisions, but we do flag when things look anomalous.
>
> <img alt="Screenshot 2021-06-11 at 20 13 35" width="924" src="https://user-images.githubusercontent.com/4483/121738566-472d9c80-caf2-11eb-82cb-a7a2bc20f68e.png">

Dear @arfon, many thanks for asking concerning the authorship order. Yes, there are some reasons that we want to explain briefly. The Sync Toolbox is a long-term project, which beginnings go back to 2006. The first synchronization procedures were developed by the first author @meinardmueller, resulting in internal research source code. This code has been successively expanded by various people from the first author's research group, resulting in an internal MATLAB toolbox. The internal toolbox was now re-implemented in Python mainly by @yiitozer and @ahnonay The JOSS paper itself was written mainly by @meinardmueller. Regardless of the respective contributions, the code and the paper's markdown was pushed to github by @ahnonay and @yiitozer; we did not care about github commit numbers or any other metrics. Internally, we discussed and agreed on the author order as stated in the paper. We want to apologize if we have deviated from git conventions, and we hope that these explanations clarify your question.
