### 2021-06-14 #3354 @faroit
👋  @magdalenafuentes @expectopatronum @lutzhamel - would any of you be willing to review this submission for JOSS? The JOSS review process takes place on GitHub and focuses on the software and a short paper. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under review is [Sync Toolbox: A Python Package for Efficient, Robust, and Accurate Music Synchronization ](https://github.com/meinardmueller/synctoolbox).
