### 2021-06-15 #3354 @lutzhamel
Hi,  I'd be happy to act as a reviewer if you still need one for this
project.

On Mon, Jun 14, 2021 at 8:51 AM Fabian-Robert Stöter <
***@***.***> wrote:

> 👋 @magdalenafuentes <https://github.com/magdalenafuentes>
> @expectopatronum <https://github.com/expectopatronum> @lutzhamel
> <https://github.com/lutzhamel> - would any of you be willing to review
> this submission for JOSS? The JOSS review process takes place on GitHub and
> focuses on the software and a short paper. We carry out our
> checklist-driven reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> The software under review is Sync Toolbox: A Python Package for
> Efficient, Robust, and Accurate Music Synchronization
> <https://github.com/meinardmueller/synctoolbox>.
>
> This issue is a "pre-review" issue in which reviewers are assigned. Once
> sufficient reviewers are recruited we will open a dedicated review issue
> where the review will take place.
>
