+++ paper/references.bib
@@ -51,7 +51,8 @@
 	isbn        = {978-1-4244-2354-5},
 	issn        = {1520-6149},
 	url-pdf   = {2009_EwertMuellerGrosche_HighResAudioSync_ICASSP.pdf},
-	url-details = {https://www.audiolabs-erlangen.de/resources/MIR/SyncRWC60}
+	url-details = {https://www.audiolabs-erlangen.de/resources/MIR/SyncRWC60},
+	doi = {10.1109/icassp.2009.4959972}
 }
 
 
@@ -81,7 +82,7 @@
   number   = {},
   pages    = {},
   year     = {2020, in press},
-  url-demo = {https://zenodo.org/record/3968389#.X7PwxmhKiUk}
+  doi       = {10.1145/3429743},
 }
 
 
