### 2021-06-30 #3354 @yiitozer
hello @faroit, I added the missing DOIs, for some references, no DOIs are available.
