### 2021-06-30 #3434 @faroit
👏 @lutzhamel @magdalenafuentes - This is the review thread for the paper. All of our communications will happen here from now on.

Both of you have checklists at the top of this thread (in that first comment) with the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.
