+++ tests/test_features.py
@@ -1,0 +1,143 @@
+import librosa
+import numpy as np
+
+from synctoolbox.feature.chroma import pitch_to_chroma, quantize_chroma, quantized_chroma_to_CENS
+from synctoolbox.feature.csv_tools import read_csv_to_df, df_to_pitch_features, df_to_pitch_onset_features
+from synctoolbox.feature.dlnco import pitch_onset_features_to_DLNCO
+from synctoolbox.feature.pitch import audio_to_pitch_features
+from synctoolbox.feature.pitch_onset import audio_to_pitch_onset_features
+from synctoolbox.feature.utils import estimate_tuning
+from utils import dict_allclose, load_dict
+
+
+def test_tuning():
+    import libfmp.c2  # TODO: This should be removed after the new release of libfmp
+    audio_1, _ = librosa.load('data_music/Schubert_D911-03_HU33.wav', 22050)
+    audio_2, _ = librosa.load('data_music/Schubert_D911-03_SC06.wav', 22050)
+    tuning_offset_1 = estimate_tuning(audio_1, 22050)
+    tuning_offset_2 = estimate_tuning(audio_2, 22050)
+
+    assert tuning_offset_1 == 7
+    assert tuning_offset_2 == 1
+
+
+def test_chroma_features():
+    f_pitch_1_gt = np.load('tests/data/f_pitch_1.npy')
+    f_pitch_2_gt = np.load('tests/data/f_pitch_2.npy')
+    f_chroma_1_gt = np.load('tests/data/f_chroma_1.npy')
+    f_chroma_2_gt = np.load('tests/data/f_chroma_2.npy')
+
+    f_chroma_1 = pitch_to_chroma(f_pitch=f_pitch_1_gt)
+    f_chroma_2 = pitch_to_chroma(f_pitch=f_pitch_2_gt)
+
+    assert np.allclose(f_chroma_1, f_chroma_1_gt, atol=1e-5)
+    assert np.allclose(f_chroma_2, f_chroma_2_gt, atol=1e-5)
+
+
+def test_quantized_chroma_features():
+    f_chroma_1_gt = np.load('tests/data/f_chroma_1.npy')
+    f_chroma_2_gt = np.load('tests/data/f_chroma_2.npy')
+    f_chroma_quantized_1_gt = np.load('tests/data/f_chroma_quantized_1.npy')
+    f_chroma_quantized_2_gt = np.load('tests/data/f_chroma_quantized_2.npy')
+
+    f_chroma_quantized_1 = quantize_chroma(f_chroma=f_chroma_1_gt)
+    f_chroma_quantized_2 = quantize_chroma(f_chroma=f_chroma_2_gt)
+
+    assert np.allclose(f_chroma_quantized_1_gt, f_chroma_quantized_1, atol=1e-5)
+    assert np.allclose(f_chroma_quantized_2_gt, f_chroma_quantized_2, atol=1e-5)
+
+
+def test_CENS_features():
+    f_chroma_quantized_1_gt = np.load('tests/data/f_chroma_quantized_1.npy')
+    f_chroma_quantized_2_gt = np.load('tests/data/f_chroma_quantized_2.npy')
+    f_CENS_1_gt = np.load('tests/data/f_CENS_1.npy')
+    f_CENS_2_gt = np.load('tests/data/f_CENS_2.npy')
+
+    f_cens_1 = quantized_chroma_to_CENS(f_chroma_quantized_1_gt, 201, 50, 50)[0]
+    f_cens_2 = quantized_chroma_to_CENS(f_chroma_quantized_2_gt, 201, 50, 50)[0]
+
[…]
