### 2021-07-14 #3434 @lutzhamel
@yiitozer I am trying to run your package in the Amazon Sagemaker/Jupyter Notebook environment and I get the following error when I try to run your simple example notebook,
```
<ipython-input-1-e4229c4ff563> in <module>
      1 # Loading some modules and defining some constants used later
      2 import time
----> 3 import librosa.display
      4 import matplotlib.pyplot as plt
      5 import IPython.display as ipd

ModuleNotFoundError: No module named 'librosa'
```
Please advise.
