### 2021-07-14 #3434 @lutzhamel
In that case I think you need to rewrite your installation instructions as follows making sure that a user must use pip to install.  Your wording implies that pip is optional to the cloning...

You **must** install the Sync Toolbox using the Python package manager pip **so that all dependencies are resolved**:

```
pip install synctoolbox
```
In order to run the example notebooks, you may clone this repository using
```
git clone https://github.com/meinardmueller/synctoolbox.git
```
