### 2021-07-14 #3434 @lutzhamel
I tried to run
```
!pip3 install synctoolbox
```
and I get the following error in the Sagemaker Notebook:
```
ERROR: Cannot uninstall 'llvmlite'. It is a distutils installed project and thus we cannot accurately determine which files belong to it which would lead to only a partial uninstall.
```
Given that your dependency script very aggressively rewrites/uninstalls many packages I cannot install this on my personal machine for testing and risk disruption of my environment.  Please suggest alternatives for me to test your software.
