### 2021-07-14 #3434 @ahnonay
Hi!
Sorry to hear you are having trouble with testing the package. It was certainly not our intention to aggressively rewrite your environment.
I have to admit, I am not familiar with Sagemaker and can not provide advice for that.
You may consider installing the synctoolbox package inside a fresh conda/virtualenv environment so that it does not interfere with the rest of your setup.
For example, with anaconda/miniconda you could write
```
conda create --name synctoolboxtest python=3.8
conda activate synctoolboxtest
pip install synctoolbox
```
or by using virtual envs:
```
python -m venv synctoolboxtest
source synctoolboxtest/bin/activate
pip install synctoolbox
```
Thank you very much for providing suggestions for improving the Readme, we will update this soon.
