### 2021-07-14 #3434 @faroit
@lutzhamel great to see that you take the functionality test seriously. To keep this compact, I would encourage you to submit issues directly in the repository. When doing so, please mention https://github.com/openjournals/joss-reviews/issues/3434 so that a link is created to this thread
