### 2021-07-14 #3434 @lutzhamel
@ahnonay @yiitozer I have now spent about 2.5 hours trying to get this software to run. First on SageMaker Notebooks and then on vanilla Ubuntu 20.04/Jupyter Notebooks.  Neither worked.  On Sagemaker I could not install because the synctoolbox package tried to uninstall the llvmlite package which is not allowed in the sagemaker environment.  On vanilla Ubuntu/Jupyter Notebooks I get the following error after doing a `!pip3 install synctoolbox`:
```
OSError                                   Traceback (most recent call last)
/tmp/ipykernel_8082/3351849963.py in <module>
      1 # Loading some modules and defining some constants used later
      2 import time
----> 3 import librosa.display
      4 import matplotlib.pyplot as plt
      5 import IPython.display as ipd

~/.local/lib/python3.8/site-packages/librosa/__init__.py in <module>
    209 # And all the librosa sub-modules
    210 from ._cache import cache
--> 211 from . import core
    212 from . import beat
    213 from . import decompose

~/.local/lib/python3.8/site-packages/librosa/core/__init__.py in <module>
      4
      5 from .convert import *  # pylint: disable=wildcard-import
----> 6 from .audio import *  # pylint: disable=wildcard-import
      7 from .spectrum import *  # pylint: disable=wildcard-import
      8 from .pitch import *  # pylint: disable=wildcard-import

~/.local/lib/python3.8/site-packages/librosa/core/audio.py in <module>
      6 import warnings
      7
----> 8 import soundfile as sf
      9 import audioread
     10 import numpy as np

~/.local/lib/python3.8/site-packages/soundfile.py in <module>
    140     _libname = _find_library('sndfile')
    141     if _libname is None:
--> 142         raise OSError('sndfile library not found')
    143     _snd = _ffi.dlopen(_libname)
    144 except OSError:

OSError: sndfile library not found
```
Please ensure that your package works on all major platforms.

 @faroit Please advise on how to proceed from here.
