### 2021-07-14 #3434 @faroit
> @faroit I am new to this, which repository should I put the synctoolbox issues in? And how do I mention an issue number for this thread from another repository if that is what is needed?

open an issue here: https://github.com/meinardmueller/synctoolbox and mention `https://github.com/openjournals/joss-reviews/issues/3434` anywhere in the text
