### 2021-07-15 8270955
Added rubberband-cli to apt.txt
