+++ apt.txt
@@ -1,1 +1,2 @@
 libsndfile1
+rubberband-cli
