### 2021-07-15 meinardmueller/synctoolbox#3 @ahnonay
Issues with soundfile when installing Sync Toolbox 
This issue arised for @lutzhamel during the JOSS review process, see also:
https://github.com/openjournals/joss-reviews/issues/3434

In some cases when installing synctoolbox via pip, one gets errors related to ``soundfile`` which is a dependency of ``librosa``.

We hope to have addressed this issue in the latest commits on the master branch in two ways:
- The example notebooks can now be run inside the browser using Binder, no installation necessary.
- We have added additional explanations for users who want to install the toolbox locally. In particular, the issues with soundfile can usually be solved by running ``sudo apt install libsndfile1``

Dear @lutzhamel, thanks for making us aware of this problem. Please let us know if the latest changes have solved this for you.
