### 2021-07-15 #3434 @ahnonay
I have created an issue for this in the Sync Toolbox repository and referenced this issue, see above.

@lutzhamel : We are really sorry this has caused you so much headache. It must have been very frustrating for you. Please let us know inside the newly created issue whether our changes have fixed this problem for you.
