### 2021-07-15 meinardmueller/synctoolbox#5 @faroit
Enable continous integration
to solve installation issues it could help to include a number of continuous integration tests that run

* runs the unit tests and installs the package
* test the package installation in a conda environment using different python versions

https://github.com/openjournals/joss-reviews/issues/3434
