### 2021-07-15 meinardmueller/synctoolbox#6 @faroit
Enable support for python 3.9
is there a specific reason why python 3.9 is not supported?

https://github.com/openjournals/joss-reviews/issues/3434
