### 2021-07-15 meinardmueller/synctoolbox#7 @faroit
Unit tests fail
running the unit tests on python 3.8 using the proposed procedure seem to fail:

```
base38 ❯ py.test tests                                                                                                (base38)
===================================================== test session starts =====================================================
platform darwin -- Python 3.8.10, pytest-6.2.4, py-1.10.0, pluggy-0.13.1
rootdir: /Users/faro/repositories/synctoolbox
collected 13 items

tests/test_dtw.py ....                                                                                                  [ 30%]
tests/test_features.py ....FF...                                                                                        [100%]

========================================================== FAILURES ===========================================================
_____________________________________________________ test_pitch_features _____________________________________________________

    def test_pitch_features():
        audio_1, _ = librosa.load('data_music/Schubert_D911-03_HU33.wav', 22050)
        audio_2, _ = librosa.load('data_music/Schubert_D911-03_SC06.wav', 22050)
        f_pitch_1_gt = np.load('tests/data/f_pitch_1.npy')
        f_pitch_2_gt = np.load('tests/data/f_pitch_2.npy')

        f_pitch_1 = audio_to_pitch_features(f_audio=audio_1,
                                            Fs=22050,
                                            tuning_offset=1,
                                            feature_rate=50,
                                            verbose=False)

        f_pitch_2 = audio_to_pitch_features(f_audio=audio_2,
                                            Fs=22050,
                                            tuning_offset=7,
                                            feature_rate=50,
                                            verbose=False)

>       assert np.allclose(f_pitch_1, f_pitch_1_gt, atol=1e-5)
E       assert False
E        +  where False = <function allclose at 0x7f85d8b23670>(array([[0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       ...,\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.]]), array([[0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       ...,\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.]]), atol=1e-05)
E        +    where <function allclose at 0x7f85d8b23670> = np.allclose

tests/test_features.py:81: AssertionError
---------------------------------------------------- Captured stdout call -----------------------------------------------------
................................................................................................................................................................................
__________________________________________________ test_pitch_onset_features __________________________________________________

    def test_pitch_onset_features():
        audio_1, _ = librosa.load('data_music/Schubert_D911-03_HU33.wav', 22050)
        audio_2, _ = librosa.load('data_music/Schubert_D911-03_SC06.wav', 22050)
        f_pitch_onset_1_gt = load_dict('tests/data/f_pitch_onset_1.pickle')
        f_pitch_onset_2_gt = load_dict('tests/data/f_pitch_onset_2.pickle')

        f_pitch_onset_1 = audio_to_pitch_onset_features(f_audio=audio_1,
                                                        Fs=22050,
                                                        tuning_offset=1,
                                                        verbose=False)

        f_pitch_onset_2 = audio_to_pitch_onset_features(f_audio=audio_2,
                                                        Fs=22050,
                                                        tuning_offset=7,
                                                        verbose=False)

>       dict_allclose(f_pitch_onset_1, f_pitch_onset_1_gt, atol=1e-5)

tests/test_features.py:101:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
tests/utils.py:26: in dict_allclose
    assert np.allclose(dict1[k], dict2[k], atol)
<__array_function__ internals>:5: in allclose
    ???
/opt/homebrew/Caskroom/miniconda/base/envs/base38/lib/python3.8/site-packages/numpy/core/numeric.py:2171: in allclose
    res = all(isclose(a, b, rtol=rtol, atol=atol, equal_nan=equal_nan))
<__array_function__ internals>:5: in isclose
    ???
/opt/homebrew/Caskroom/miniconda/base/envs/base38/lib/python3.8/site-packages/numpy/core/numeric.py:2272: in isclose
    return within_tol(x, y, atol, rtol)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

x = array([[[3.10917800e+04],
        [3.11484694e+04],
        [3.11824830e+04],
        [3.12164966e+04],
        [3.123...9819230e-09],
        [1.69793182e-09],
        [1.56447331e-09],
        [1.20539503e-09],
        [9.61932239e-10]]])
y = array([[[3.10917800e+04],
        [3.11484694e+04],
        [3.11824830e+04],
        [3.12164966e+04],
        [3.123...7001423e-09],
        [1.67276090e-09],
        [1.54219300e-09],
        [1.18766368e-09],
        [9.48867734e-10]]])
atol = 1e-08, rtol = 1e-05

    def within_tol(x, y, atol, rtol):
        with errstate(invalid='ignore'):
>           return less_equal(abs(x-y), atol + rtol * abs(y))
E           ValueError: operands could not be broadcast together with shapes (2,138,1) (2,136,1)

/opt/homebrew/Caskroom/miniconda/base/envs/base38/lib/python3.8/site-packages/numpy/core/numeric.py:2258: ValueError
---------------------------------------------------- Captured stdout call -----------------------------------------------------
................................................................................................................................................................................21 (2, 138, 1) (2, 136, 1)
=================================================== short test summary info ===================================================
FAILED tests/test_features.py::test_pitch_features - assert False
FAILED tests/test_features.py::test_pitch_onset_features - ValueError: operands could not be broadcast together with shapes ...
================================================ 2 failed, 11 passed in 56.14s ================================================
```

related #5
https://github.com/openjournals/joss-reviews/issues/3434
