### 2021-07-15 meinardmueller/synctoolbox#3 @lutzhamel
Installing `libsndfile1` on Ubuntu 20.04 did the trick.  I was able to run the `sync_audio_audio_simple.ipynb` and the `sync_audio_score_full.ipynb` locally on the linux system.
