### 2021-07-15 #3434 @lutzhamel
@ahnonay No worries.  I am just doing my job as a reviewer.  I am glad that you were able to resolve those issues so quickly.  I was able to run two of the three notebooks locally on my Linux machine.

The third notebook, `sync_audio_audio_full.ipynb`, I was not able to fully run due to the rubberband dependency.  Unfortunately this software does not come as an apt package for Linux and I started to go down the rabbit hole trying to build this thing from scratch as recommended by their poorly written documentation.  I gave up after chasing this installation for about 30 minutes.

Question:  what does this notebook illustrate that the other two notebooks do not illustrate? If the delta between the other two notebooks and this one is small I would suggest to remove it as a demo because the installation requirement of the rubberband lib is a substantial burden due to its poorly written documentation and does not cast a good light on your software package.
