### 2021-07-16 #3434 @ahnonay
Hi Lutzhamel,
glad to hear you were able to run at least most of the notebooks now :)!

Regarding the rubberband dependency: We are quite unhappy with this as well. In fact, we are currently creating a Python package that can replace and improve upon rubberband (https://github.com/meinardmueller/libtsm), but this won't be ready for some more time. The application scenario that we need rubberband for is often requested by other researchers in our area, so we would like to keep this functionality and showcase it in one of the notebooks. If rubberband is not available, at least the other cells in the notebook can be run.

Since you are running Ubuntu, you should be able to install rubberband using:
``sudo apt install rubberband-cli``
I'll add this as a note in the ``sync_audio_audio_full.ipynb`` notebook.
