### 2021-07-16 meinardmueller/synctoolbox#3 @ahnonay
Happy to hear that! I'll close this issue for now. Please reopen, if you think this is not yet resolved.
Regarding the remaining problems related to rubberband, please see my response in
openjournals/joss-reviews#3434
