### 2021-07-16 #3434 @lutzhamel
@ahnonay I am happy to say that I was able to run the final notebook with the apt rubberband package.

@faroit My review is complete and I recommend publication of this package/paper.
