### 2021-07-19 1a727e9
Updated minimum libfmp version to address issue #4, enabled support for Python 3.9 to fix issue #6
