+++ setup.py
@@ -34,8 +34,8 @@ setup(
                       'pretty_midi >= 0.2.0, < 1.0.0',
                       'pysoundfile >= 0.9.0, < 1.0.0',
                       'scipy >= 1.3.0, < 2.0.0',
-                      'libfmp >= 1.1.0, < 2.0.0'],
-    python_requires='>=3.6, <3.9',
+                      'libfmp >= 1.2.0, < 2.0.0'],
+    python_requires='>=3.6, <4.0',
     extras_require={
         'tests': ['pytest == 6.2.*'],
         'docs': ['sphinx == 4.0.*',
