### 2021-07-19 meinardmueller/synctoolbox#6 @ahnonay
The package works with python 3.9 as well, so I've updated the setup.py in this regard. Closing this for now, please reopen if you think this has not been addressed yet.
