### 2021-07-26 d98515d
scipy version requirement fix
