### 2021-07-26 e22ed46
numba >= 0.51.0 is compatible with python > =3.7
