### 2021-07-26 meinardmueller/synctoolbox#7 @yiitozer
Dear @faroit, thank you for pointing this out.

I fixed this problem, it was due to the scipy version. Could you try it out again?
