### 2021-07-26 meinardmueller/synctoolbox#7 @faroit
thanks @yiitozer. I think I would leave this open until #5 is addressed so I wouldn't have to try it manually
