### 2021-07-27 meinardmueller/synctoolbox#5 @yiitozer
I created a workflow which runs the unit tests in a conda environment using python 3.8 version:
https://github.com/meinardmueller/synctoolbox/actions/runs/1070732431

Would it be necessary to have two more workflows for python 3.7 and 3.9 versions?
