### 2021-07-27 meinardmueller/synctoolbox#5 @faroit
great!

> Would it be necessary to have two more workflows for python 3.7 and 3.9 versions?

i usually test multiple python version for the pip tests and only a single one conda test because the conda workflows aren't that quick and do not cache that well..

See [this example](https://github.com/sigsep/open-unmix-pytorch/blob/master/.github/workflows/test_cli.yml) how run matrix builds with different python version using a single workflow
