+++ .github/workflows/python-package.yml
@@ -28,6 +28,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install flake8 pytest pysndfile
+        apt-get install libsndfile1
         python -m pip install .
     - name: Lint with flake8
       run: |
