### 2021-07-27 c119050
Update and rename python-package.yml to test_pip.yml
