+++ .github/workflows/test_pip.yml
@@ -24,6 +24,10 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: ${{ matrix.python-version }}
+    - name: Install libnsdfile
+      run: |
+        sudo apt update
+        sudo apt install libsndfile1-dev libsndfile1    
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
