+++ .github/workflows/test_pip.yml
@@ -33,6 +33,9 @@ jobs:
         python -m pip install --upgrade pip
         python -m pip install flake8 pytest sndfile
         python -m pip install .
+        python --version
+        pip --version
+        python -m pip list
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
