### 2021-07-27 dd62228
Update test_conda.yml
pip install update
