+++ .github/workflows/test_conda.yml
@@ -21,7 +21,7 @@ jobs:
     - name: Install dependencies
       run: |
         conda install -c conda-forge libsndfile
-        pip install .
+        python -m pip install -e .['tests']
     - name: Lint with flake8
       run: |
         conda install flake8
@@ -31,5 +31,4 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        conda install pytest
         pytest
