### 2021-07-27 8b73129
Update test_pip.yml
pip install update
