+++ .github/workflows/test_pip.yml
@@ -31,8 +31,8 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install flake8 pytest sndfile
-        python -m pip install .
+        python -m pip install flake8
+        python -m pip install -e .['tests']
         python --version
         pip --version
         python -m pip list
