### 2021-07-27 meinardmueller/synctoolbox#5 @yiitozer
Thank you, the example has been very helpful.

I created two pull requests following your instructions:
https://github.com/meinardmueller/synctoolbox/pulls

I will merge them if you confirm it.
