### 2021-07-27 meinardmueller/synctoolbox#5 @faroit
@yiitozer both look good to me.

Just noted: there are a bunch of deprecations shown in the unit tests when running with numpy 1.20 (See [here](https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations)). I would expect this package would break with a future numpy version soon, so better to fix this sooner rather than later
