### 2021-07-27 meinardmueller/synctoolbox#5 @yiitozer
@faroit, thank you for pointing it out. I fixed it!
