### 2021-08-04 meinardmueller/synctoolbox#11 @magdalenafuentes
Nits on docs and repo
Hey @ahnonay and @yiitozer thanks for creating this useful toolbox. Looks really good. I'll drop here some nits to follow JOSS guidelines and make the repo even clearer:

- JOSS recommends to add a statement of need in the documentation as well, so maybe it's worth to add not only the API description in the docs, but also a brief description and the statement in the index.
- I recommend you add a brief description in the `about` field of the GitHub repo as well as a link to the `docs` in the `website` field. It will make it quicker for people to get to the docs and also understand what the repo is about

openjournals/joss-reviews#3434
