### 2021-08-04 meinardmueller/synctoolbox#12 @magdalenafuentes
Release with Python 3.9
Just as a reminder, since the release in pip right now is not working with `python 3.9`. In a conda environment with that python version I'm getting:

```
(joss_review) cusp-vmn244:repos mf3734$ pip install synctoolbox
ERROR: Could not find a version that satisfies the requirement synctoolbox (from versions: none)
ERROR: No matching distribution found for synctoolbox
```

Related to #6

openjournals/joss-reviews#3434
