### 2021-08-04 meinardmueller/synctoolbox#13 @magdalenafuentes
I saw that you discussed about it [here](https://github.com/openjournals/joss-reviews/issues/3434) but I don't have sudo status in this machine, neither in the linux ones I was planning to test the software on. I'm happy to move on and close this issue if you think there's no workaround in the near future since you already added a comment in the notebook explaining this. I just want to make sure that the error I'm getting is the same you expect and there's nothing else going on.

I confirm that the other cells in this notebook work fine for me.
