### 2021-08-05 e64d363
a statement of need added into the documentation. (issue #11)
