### 2021-08-05 meinardmueller/synctoolbox#13 @yiitozer
We're planning on replacing the rubberband package with libtsm soon:
https://github.com/meinardmueller/libtsm

once we ensure that the libtsm package is stable. Hopefully, that's OK to keep the notebook as it is for now.
