### 2021-08-05 meinardmueller/synctoolbox#7 @faroit
@yiitozer i guess this can be closed? Maybe you could add github action unit test status badges to the main readme?
