### 2021-08-05 meinardmueller/synctoolbox#7 @yiitozer
Done! I'm closing this issue.
