### 2021-08-05 meinardmueller/synctoolbox#13 @magdalenafuentes
Ok, makes sense. Will close this then. Thanks!
