### 2021-08-05 meinardmueller/synctoolbox#7 @yiitozer
I found out that the scipy version was the cause of this issue. Could you tell me what you scipy version is? It should be greater than or equal to 1.7.0.

We'll update the package on the PyPI in short:
https://pypi.org/project/synctoolbox/
