### 2021-08-05 meinardmueller/synctoolbox#7 @magdalenafuentes
I have `scipy 1.7.1`, here my full conda environment:

```
Package             Version
------------------- -------------------
appdirs             1.4.4
appnope             0.1.2
argon2-cffi         20.1.0
async-generator     1.10
attrs               21.2.0
audioread           2.1.9
backcall            0.2.0
bleach              4.0.0
certifi             2021.5.30
cffi                1.14.6
charset-normalizer  2.0.4
cycler              0.10.0
debugpy             1.4.1
decorator           5.0.9
defusedxml          0.7.1
entrypoints         0.3
idna                3.2
iniconfig           1.1.1
ipykernel           6.0.3
ipython             7.26.0
ipython-genutils    0.2.0
ipywidgets          7.6.3
jedi                0.18.0
Jinja2              3.0.1
joblib              1.0.1
jsonschema          3.2.0
jupyter             1.0.0
jupyter-client      6.1.12
jupyter-console     6.4.0
jupyter-core        4.7.1
jupyterlab-pygments 0.1.2
jupyterlab-widgets  1.0.0
kiwisolver          1.3.1
libfmp              1.2.1
librosa             0.8.1
llvmlite            0.36.0
MarkupSafe          2.0.1
matplotlib          3.4.2
matplotlib-inline   0.1.2
mido                1.2.10
mistune             0.8.4
music21             5.7.2
nbclient            0.5.3
nbconvert           6.1.0
nbformat            5.1.3
nest-asyncio        1.5.1
notebook            6.4.0
numba               0.53.1
numpy               1.21.1
packaging           21.0
pandas              1.3.1
pandocfilters       1.4.3
parso               0.8.2
pexpect             4.8.0
pickleshare         0.7.5
Pillow              8.3.1
pip                 21.2.2
pluggy              0.13.1
pooch               1.4.0
pretty-midi         0.2.9
prometheus-client   0.11.0
prompt-toolkit      3.0.19
ptyprocess          0.7.0
py                  1.10.0
pycparser           2.20
Pygments            2.9.0
pyparsing           2.4.7
pyrsistent          0.18.0
pyrubberband        0.3.0
PySoundFile         0.9.0.post1
pytest              6.2.4
python-dateutil     2.8.2
pytz                2021.1
pyzmq               22.2.0
qtconsole           5.1.1
QtPy                1.9.0
requests            2.26.0
resampy             0.2.2
scikit-learn        0.24.2
scipy               1.7.1
Send2Trash          1.7.1
setuptools          52.0.0.post20210125
six                 1.16.0
SoundFile           0.10.3.post1
synctoolbox         1.0.0
terminado           0.10.1
testpath            0.5.0
threadpoolctl       2.2.0
toml                0.10.2
tornado             6.1
traitlets           5.0.5
urllib3             1.26.6
wcwidth             0.2.5
webencodings        0.5.1
wheel               0.36.2
widgetsnbextension  3.5.1
```
