### 2021-08-05 meinardmueller/synctoolbox#7 @magdalenafuentes
@yiitozer I tried with `scipy==1.7.0` as well and I still see the errors. Let me know how to proceed.
