### 2021-08-05 2a77154
Fix unit tests
- Unit tests added for the filterbank and peak search
- test_conda.yml fixed (python version)
- scipy.signal.filtfilt function replaced with scipy.signal.sosfiltfilt due to numerical instabilities.
