### 2021-08-05 meinardmueller/synctoolbox#7 @magdalenafuentes
Yes, it's the same:

```
tests/test_dtw.py ....                                                                                                       [ 30%]
tests/test_features.py ....FF...                                                                                             [100%]

============================================================= FAILURES =============================================================
_______________________________________________________ test_pitch_features ________________________________________________________

    def test_pitch_features():
        audio_1, _ = librosa.load('data_music/Schubert_D911-03_HU33.wav', 22050)
        audio_2, _ = librosa.load('data_music/Schubert_D911-03_SC06.wav', 22050)
        f_pitch_1_gt = np.load('tests/data/f_pitch_1.npy')
        f_pitch_2_gt = np.load('tests/data/f_pitch_2.npy')

        f_pitch_1 = audio_to_pitch_features(f_audio=audio_1,
                                            Fs=22050,
                                            tuning_offset=1,
                                            feature_rate=50,
                                            verbose=False)

        f_pitch_2 = audio_to_pitch_features(f_audio=audio_2,
                                            Fs=22050,
                                            tuning_offset=7,
                                            feature_rate=50,
                                            verbose=False)

        assert np.allclose(f_pitch_1, f_pitch_1_gt, atol=1e-5)
>       assert np.allclose(f_pitch_2, f_pitch_2_gt, atol=1e-5)
E       assert False
E        +  where False = <function allclose at 0x7fc7c81ea430>(array([[0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       ...,\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.]]), array([[0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       ...,\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.]]), atol=1e-05)
E        +    where <function allclose at 0x7fc7c81ea430> = np.allclose

tests/test_features.py:82: AssertionError
------------------------------------------------------- Captured stdout call -------------------------------------------------------
................................................................................................................................................................................
____________________________________________________ test_pitch_onset_features _____________________________________________________

    def test_pitch_onset_features():
        audio_1, _ = librosa.load('data_music/Schubert_D911-03_HU33.wav', 22050)
        audio_2, _ = librosa.load('data_music/Schubert_D911-03_SC06.wav', 22050)
        f_pitch_onset_1_gt = load_dict('tests/data/f_pitch_onset_1.pickle')
        f_pitch_onset_2_gt = load_dict('tests/data/f_pitch_onset_2.pickle')

        f_pitch_onset_1 = audio_to_pitch_onset_features(f_audio=audio_1,
                                                        Fs=22050,
                                                        tuning_offset=1,
                                                        verbose=False)

        f_pitch_onset_2 = audio_to_pitch_onset_features(f_audio=audio_2,
                                                        Fs=22050,
                                                        tuning_offset=7,
                                                        verbose=False)

>       dict_allclose(f_pitch_onset_1, f_pitch_onset_1_gt, atol=1e-5)

tests/test_features.py:101:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
tests/utils.py:26: in dict_allclose
    assert np.allclose(dict1[k], dict2[k], atol)
<__array_function__ internals>:5: in allclose
    ???
/Users/mf3734/miniconda3/envs/joss_review_2/lib/python3.8/site-packages/numpy/core/numeric.py:2249: in allclose
    res = all(isclose(a, b, rtol=rtol, atol=atol, equal_nan=equal_nan))
<__array_function__ internals>:5: in isclose
    ???
/Users/mf3734/miniconda3/envs/joss_review_2/lib/python3.8/site-packages/numpy/core/numeric.py:2358: in isclose
    return within_tol(x, y, atol, rtol)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

x = array([[[3.10917800e+04],
        [3.11484694e+04],
        [3.11824830e+04],
        [3.12164966e+04],
        [3.123...0950938e-09],
        [1.70858722e-09],
        [1.57374780e-09],
        [1.21270285e-09],
        [9.67767082e-10]]])
y = array([[[3.10917800e+04],
        [3.11484694e+04],
        [3.11824830e+04],
        [3.12164966e+04],
        [3.123...8774396e-09],
        [1.68822468e-09],
        [1.55594432e-09],
        [1.19865935e-09],
        [9.56583904e-10]]])
atol = 1e-08, rtol = 1e-05

    def within_tol(x, y, atol, rtol):
        with errstate(invalid='ignore'):
>           return less_equal(abs(x-y), atol + rtol * abs(y))
E           ValueError: operands could not be broadcast together with shapes (2,138,1) (2,136,1)

/Users/mf3734/miniconda3/envs/joss_review_2/lib/python3.8/site-packages/numpy/core/numeric.py:2339: ValueError
------------------------------------------------------- Captured stdout call -------------------------------------------------------
................................................................................................................................................................................21 (2, 138, 1) (2, 136, 1)
```

Then a lot of warnings I skip for clarity and

```
-- Docs: https://docs.pytest.org/en/stable/warnings.html
===================================================== short test summary info ======================================================
FAILED tests/test_features.py::test_pitch_features - assert False
FAILED tests/test_features.py::test_pitch_onset_features - ValueError: operands could not be broadcast together with shapes (2,13...
=========================================== 2 failed, 11 passed, 1482 warnings in 42.34s ===========================================
```
