### 2021-08-05 meinardmueller/synctoolbox#7 @yiitozer
@magdalenafuentes thank you, I could reproduce it. I'll work on it.
