### 2021-08-06 meinardmueller/synctoolbox#7 @yiitozer
@magdalenafuentes, I noticed that the problem is related to the filtfilt function (scipy.signal.filtfilt).

https://github.com/meinardmueller/synctoolbox/runs/3261974717?check_suite_focus=true

Somehow its result is not deterministic, although the scipy version is the same for each python version.
