### 2021-08-06 meinardmueller/synctoolbox#7 @yiitozer
I solved the issue with the pull request #15.

@magdalenafuentes, can you please run the tests with the up-to-date version of the master branch?
If that also works on your local machine, I will create a new release in pip. This would close issue #12 as well.
