### 2021-08-06 meinardmueller/synctoolbox#7 @magdalenafuentes
Hey @yiitozer I pulled the latest changes and this is what I'm getting when running the testes (`pytest test`):

```
(joss_review_2) cusp-vmn244:synctoolbox mf3734$ git pull origin master
remote: Enumerating objects: 73, done.
remote: Total 48 (delta 24), reused 39 (delta 20), pack-reused 0
Unpacking objects: 100% (48/48), done.
From https://github.com/meinardmueller/synctoolbox
 * branch            master     -> FETCH_HEAD
   cb52c9c..e5fba5c  master     -> origin/master
Updating cb52c9c..e5fba5c
Fast-forward
 .github/workflows/test_conda.yml    |   8 ++++++--
 synctoolbox/feature/filterbank.py   |  12 +-----------
 synctoolbox/feature/pitch.py        |   7 +++----
 synctoolbox/feature/pitch_onset.py  |  15 ++++++++-------
 tests/data/f_CENS_1.npy             | Bin 14624 -> 14624 bytes
 tests/data/f_DLNCO_1.npy            | Bin 721856 -> 721856 bytes
 tests/data/f_DLNCO_2.npy            | Bin 658688 -> 658688 bytes
 tests/data/f_chroma_1.npy           | Bin 721856 -> 721856 bytes
 tests/data/f_chroma_2.npy           | Bin 658688 -> 658688 bytes
 tests/data/f_chroma_quantized_1.npy | Bin 721856 -> 721856 bytes
 tests/data/f_filtfilt.npy           | Bin 0 -> 1060832 bytes
 tests/data/f_onset.npy              | Bin 0 -> 106184 bytes
 tests/data/f_pitch_1.npy            | Bin 7698560 -> 7698560 bytes
 tests/data/f_pitch_2.npy            | Bin 7024768 -> 7024768 bytes
 tests/data/f_pitch_onset_1.pickle   | Bin 160542 -> 160558 bytes
 tests/data/f_pitch_onset_2.pickle   | Bin 168946 -> 168562 bytes
 tests/data/fb.pickle                | Bin 0 -> 22817 bytes
 tests/data/peaks.npy                | Bin 0 -> 1008 bytes
 tests/data/thresh.npy               | Bin 0 -> 106184 bytes
 tests/data/wp_high_res.npy          | Bin 125616 -> 125600 bytes
 tests/data/wp_simple.npy            | Bin 125072 -> 125072 bytes
 tests/test_features.py              |  19 +++++++++++++++++--
 tests/utils.py                      |  25 +++++++++++++++++++++----
 23 files changed, 56 insertions(+), 30 deletions(-)
 create mode 100644 tests/data/f_filtfilt.npy
 create mode 100644 tests/data/f_onset.npy
 create mode 100644 tests/data/fb.pickle
 create mode 100644 tests/data/peaks.npy
 create mode 100644 tests/data/thresh.npy
(joss_review_2) cusp-vmn244:synctoolbox mf3734$ pytest tests
======================================================= test session starts ========================================================
platform darwin -- Python 3.8.11, pytest-6.2.4, py-1.10.0, pluggy-0.13.1
rootdir: /Users/mf3734/Documents/repos/review/2021/JOSS/synctoolbox
collected 15 items

tests/test_dtw.py ....                                                                                                       [ 26%]
tests/test_features.py ....F.FF...                                                                                           [100%]

============================================================= FAILURES =============================================================
_________________________________________________________ test_filterbank __________________________________________________________

    def test_filterbank():
        fb_gt = load_dict('tests/data/fb.pickle')
        fb = generate_filterbank(semitone_offset_cents=7)
>       filterbank_equal(fb, fb_gt)

tests/test_features.py:67:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

fb = {21: {'a': array([  1.        ,  -7.8384458 ,  27.03303336, -53.5719055 ,
        66.71929192, -53.47244784,  26.93275... -0.0245074 ,  0.08410023, -0.16623302,  0.20698032,
       -0.16623302,  0.08410023, -0.0245074 ,  0.00315005])}, ...}
fb_gt = {21: array([[ 0.00315176, -0.00615725,  0.00315176,  1.        , -1.95801479,
         0.99720306],
       [ 1.       ...     0.99886093],
       [ 1.        , -1.94980891,  1.        ,  1.        , -1.94659207,
         0.99890433]]), ...}

    def filterbank_equal(fb, fb_gt):
        """Checks whether the entries of two filterbanks are equal to each other."""
        assert fb.keys() == fb_gt.keys(), 'The MIDI indices of two filterbanks are not identical.'

        for k in fb:
>           assert np.array_equal(fb[k], fb_gt[k])
E           AssertionError

tests/utils.py:43: AssertionError
_______________________________________________________ test_pitch_features ________________________________________________________

    def test_pitch_features():
        audio_1, _ = librosa.load('data_music/Schubert_D911-03_HU33.wav', 22050)
        audio_2, _ = librosa.load('data_music/Schubert_D911-03_SC06.wav', 22050)
        f_pitch_1_gt = np.load('tests/data/f_pitch_1.npy')
        f_pitch_2_gt = np.load('tests/data/f_pitch_2.npy')

        f_pitch_1 = audio_to_pitch_features(f_audio=audio_1,
                                            Fs=22050,
                                            tuning_offset=1,
                                            feature_rate=50,
                                            verbose=False)

        f_pitch_2 = audio_to_pitch_features(f_audio=audio_2,
                                            Fs=22050,
                                            tuning_offset=7,
                                            feature_rate=50,
                                            verbose=False)

>       assert np.allclose(f_pitch_1, f_pitch_1_gt, atol=1e-5)
E       assert False
E        +  where False = <function allclose at 0x7fbc602755e0>(array([[0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       ...,\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.]]), array([[0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       ...,\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.],\n       [0., 0., 0., ..., 0., 0., 0.]]), atol=1e-05)
E        +    where <function allclose at 0x7fbc602755e0> = np.allclose

tests/test_features.py:96: AssertionError
------------------------------------------------------- Captured stdout call -------------------------------------------------------
................................................................................................................................................................................
____________________________________________________ test_pitch_onset_features _____________________________________________________

    def test_pitch_onset_features():
        audio_1, _ = librosa.load('data_music/Schubert_D911-03_HU33.wav', 22050)
        audio_2, _ = librosa.load('data_music/Schubert_D911-03_SC06.wav', 22050)
        f_pitch_onset_1_gt = load_dict('tests/data/f_pitch_onset_1.pickle')
        f_pitch_onset_2_gt = load_dict('tests/data/f_pitch_onset_2.pickle')

        f_pitch_onset_1 = audio_to_pitch_onset_features(f_audio=audio_1,
                                                        Fs=22050,
                                                        tuning_offset=1,
                                                        verbose=False)

        f_pitch_onset_2 = audio_to_pitch_onset_features(f_audio=audio_2,
                                                        Fs=22050,
                                                        tuning_offset=7,
                                                        verbose=False)

>       dict_allclose(f_pitch_onset_1, f_pitch_onset_1_gt, atol=1e-5)

tests/test_features.py:116:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

dict1 = {21: array([[[3.10917800e+04],
        [3.11484694e+04],
        [3.11824830e+04],
        [3.12164966e+04],
        [...0e-10],
        [4.25231229e-10],
        [2.19417328e-10],
        [4.31439692e-10],
        [1.28464351e-09]]]), ...}
dict2 = {21: array([[[3.10917800e+04],
        [3.11484694e+04],
        [3.11824830e+04],
        [3.12164966e+04],
        [...8e-10],
        [4.24328646e-10],
        [2.18858615e-10],
        [4.29832782e-10],
        [1.28280842e-09]]]), ...}
atol = 1e-05, rtol = 1e-05

    def dict_allclose(dict1, dict2, atol=1e-5, rtol=1e-5):
        """Checks whether the entries of two dictionaries are close to each other."""
        assert dict1.keys() == dict2.keys(), 'The midi indices of two feature arrays are not identical.'
        for k in dict1:
>           assert dict1[k].shape == dict2[k].shape, f'Features have different shapes for the midi index {k}.' \
                                                     f'The first feature array f_1[{k}] has {dict1[k].shape} elements;' \
                                                     f'The second feature array f_2[{k}] has {dict2[k].shape} elements:'
E           AssertionError: Features have different shapes for the midi index 21.The first feature array f_1[21] has (2, 138, 1) elements;The second feature array f_2[21] has (2, 136, 1) elements:

tests/utils.py:25: AssertionError

```
..warnings..
```
-- Docs: https://docs.pytest.org/en/stable/warnings.html
===================================================== short test summary info ======================================================
FAILED tests/test_features.py::test_filterbank - AssertionError
FAILED tests/test_features.py::test_pitch_features - assert False
FAILED tests/test_features.py::test_pitch_onset_features - AssertionError: Features have different shapes for the midi index 21.T...
=========================================== 3 failed, 12 passed, 1482 warnings in 40.02s ===========================================
```

So I'm getting one more error now. I'm quite confused since your tests are passing remotely. Do they run ok locally for you? Should we double check test config?
