### 2021-08-06 meinardmueller/synctoolbox#7 @yiitozer
Hi @magdalenafuentes,

I'm also quite surprised. Yes, it works on my local machine.

Maybe that's a caching problem. (I don't really think though) Can you try it without caching:

```
pytest --cache-clear
```

If it doesn't work maybe installing the repository once more can help:

```
pip install -e .[tests]
```

As you mentioned, the remote tests have also been successful, after I fixed the numeric instabilities.
