### 2021-08-06 meinardmueller/synctoolbox#7 @magdalenafuentes
Clearing the `cache` didn't work, and reinstalling the repository didn't work either but it reduced the errors to only one:

```
(joss_review_2) cusp-vmn244:synctoolbox mf3734$ pytest tests
================================================================================ test session starts ================================================================================
platform darwin -- Python 3.8.11, pytest-6.2.4, py-1.10.0, pluggy-0.13.1
rootdir: /Users/mf3734/Documents/repos/review/2021/JOSS/synctoolbox
collected 15 items

tests/test_dtw.py ....                                                                                                                                                        [ 26%]
tests/test_features.py ....F......                                                                                                                                            [100%]

===================================================================================== FAILURES ======================================================================================
__________________________________________________________________________________ test_filterbank __________________________________________________________________________________

    def test_filterbank():
        fb_gt = load_dict('tests/data/fb.pickle')
        fb = generate_filterbank(semitone_offset_cents=7)
>       filterbank_equal(fb, fb_gt)

tests/test_features.py:67:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

fb = {21: array([[ 0.00315176, -0.00615725,  0.00315176,  1.        , -1.95801479,
         0.99720306],
       [ 1.       ...     0.99886093],
       [ 1.        , -1.94980891,  1.        ,  1.        , -1.94659207,
         0.99890433]]), ...}
fb_gt = {21: array([[ 0.00315176, -0.00615725,  0.00315176,  1.        , -1.95801479,
         0.99720306],
       [ 1.       ...     0.99886093],
       [ 1.        , -1.94980891,  1.        ,  1.        , -1.94659207,
         0.99890433]]), ...}

    def filterbank_equal(fb, fb_gt):
        """Checks whether the entries of two filterbanks are equal to each other."""
        assert fb.keys() == fb_gt.keys(), 'The MIDI indices of two filterbanks are not identical.'

        for k in fb:
>           assert np.array_equal(fb[k], fb_gt[k])
E           AssertionError

tests/utils.py:43: AssertionError
================================================================================= warnings summary ==================================================================================
../../../../../../miniconda3/envs/joss_review_2/lib/python3.8/site-packages/librosa/core/constantq.py:1059
  /Users/mf3734/miniconda3/envs/joss_review_2/lib/python3.8/site-packages/librosa/core/constantq.py:1059: DeprecationWarning: `np.complex` is a deprecated alias for the builtin `complex`. To silence this warning, use `complex` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.complex128` here.
  Deprecated in NumPy 1.20; for more details and guidance: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
    dtype=np.complex,

tests/test_features.py::test_tuning
tests/test_features.py::test_tuning
  /Users/mf3734/miniconda3/envs/joss_review_2/lib/python3.8/site-packages/librosa/util/utils.py:2099: DeprecationWarning: `np.float` is a deprecated alias for the builtin `float`. To silence this warning, use `float` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.float64` here.
  Deprecated in NumPy 1.20; for more details and guidance: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
    np.dtype(np.float): np.complex,

tests/test_features.py::test_tuning
tests/test_features.py::test_tuning
  /Users/mf3734/miniconda3/envs/joss_review_2/lib/python3.8/site-packages/librosa/util/utils.py:2099: DeprecationWarning: `np.complex` is a deprecated alias for the builtin `complex`. To silence this warning, use `complex` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.complex128` here.
  Deprecated in NumPy 1.20; for more details and guidance: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
    np.dtype(np.float): np.complex,

-- Docs: https://docs.pytest.org/en/stable/warnings.html
============================================================================== short test summary info ==============================================================================
FAILED tests/test_features.py::test_filterbank - AssertionError
===================================================================== 1 failed, 14 passed, 5 warnings in 27.67s =====================================================================
```

If you change the line 43 in `tests/utils.py` to `assert np.allclose(fb[k], fb_gt[k])` instead of using `array_equal` the tests pass for me. Would that make sense for you?
