### 2021-08-09 6400b6b
array_equal -> allclose, loosen strictness of filterbank_equal function
