### 2021-08-09 meinardmueller/synctoolbox#7 @yiitozer
@magdalenafuentes , thank you for your suggestion.

Sure, I changed it to allclose, the condition doesn't have to be that strict.
