### 2021-08-09 meinardmueller/synctoolbox#12 @yiitozer
Dear @magdalenafuentes, I launched a new release. The new release is compatible with python3.9:
https://pypi.org/project/synctoolbox/1.0.1/
