### 2021-08-09 meinardmueller/synctoolbox#7 @magdalenafuentes
Hey @yiitozer did that suggestion work for you? If so I can finish my review and accept the package/paper!
