+++ paper/references.bib
@@ -75,14 +75,17 @@
 
 
 @article{WeissZAGKVM20_WinterreiseDataset_ACM-JOCCH,
-  author   = {Christof Wei{\ss} and Frank Zalkow and Vlora Arifi-M{\"u}ller and Harald Grohganz and Hendrik Vincent Koops and Anja Volk and Meinard M{\"u}ller},
-  title    = {{S}chubert {W}interreise Dataset: {A} Multimodal Scenario for Music Analysis},
-  journal  = {{ACM} Journal on Computing and Cultural Heritage ({JOCCH})},
-  volume   = {},
-  number   = {},
-  pages    = {},
-  year     = {2020, in press},
-  doi       = {10.1145/3429743},
+  author  = {Christof Wei{\ss} and Frank Zalkow and Vlora Arifi-M{\"u}ller and Meinard M{\"u}ller and Hendrik Vincent Koops and Anja Volk and Harald G. Grohganz},
+  title   = {{S}chubert {W}interreise Dataset: A Multimodal Scenario for Music Analysis},
+  journal = {{ACM} Journal on Computing and Cultural Heritage ({JOCCH})},
+  volume  = {14},
+  number  = {2},
+  pages   = {25:1--18},
+  year    = {2021},
+  doi     = {10.1145/3429743},
+  url     = {https://dl.acm.org/doi/10.1145/3429743},
+  url-pdf = {https://dl.acm.org/doi/pdf/10.1145/3429743},
+  url-details = {https://doi.org/10.5281/zenodo.5139893}
 }
 
 
