### 2021-08-10 #3434 @faroit
@yiitozer great!  At this point, could you please make a new release of the main repo that includes all of the changes that have resulted from the review? I would also suggest to issue a new pypi release at the same time, using the same tag.

Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the archive version, please make sure that:

    The title of the archive is the same as the JOSS paper title
    The authors of the archive are the same as the JOSS paper authors
