### 2021-08-10 #3434 @faroit
@yiitozer for convenience, I would also suggest to merge the paper branch into the master to match the tag of the paper with the pypi release tag
