### 2021-08-10 #3434 @faroit
> @faroit, I created a new release both on github and pypi. Is that necessary that I merge the paper branch into the master? The version is same on both the branches (1.0.2).

I guess this is fine then. @arfon ?
