### 2021-08-12 #3434 @arfon
> @faroit, I created a new release both on github and pypi. Is that necessary that I merge the paper branch into the master? The version is same on both the branches (1.0.2).

The paper doesn't need to be merged into `master`. It's completely up to you.
