+++ paper/paper.md
@@ -32,41 +32,41 @@ link-citations: yes
 ---
 
 # Summary
-Music can be described and represented in many different ways, including sheet music, symbolic representations, and audio recordings [@Mueller15_FMP_SPRINGER]. For each of these representations, there may exist different versions (e.g., recordings performed by different orchestras and conductors) that correspond to the same musical work. 
+Music can be described and represented in many different ways, including as sheet music, symbolic representations, and audio recordings [@Mueller15_FMP_SPRINGER]. For each of these representations, different versions (e.g., recordings performed by different orchestras and conductors) that correspond to the same musical work may exist. 
 Music information retrieval (MIR) aims at developing techniques and tools for organizing, understanding, and searching this information in a robust, efficient, and intelligent manner.
-In this context, various alignment and synchronization procedures have been developed with the common goal to automatically link several types of music representations, thus coordinating the multiple information sources related to a given musical work.
+In this context, various alignment and synchronization procedures have been developed with the common goal of automatically linking several types of music representations, thus coordinating the multiple information sources related to a given musical work.
 In the design and implementation of synchronization algorithms, one has to deal with a delicate tradeoff between efficiency, robustness, and accuracy---requirements leading to various approaches with many design choices.
-In this contribution, we introduce a Python package called *Sync Toolbox*, which provides open-source reference implementations for full-fledged music synchronization pipelines and yields state-of-the-art alignment results for a wide range of Western music.
+In this contribution, we introduce a Python package called *Sync Toolbox* that provides open-source reference implementations for full-fledged music synchronization pipelines and yields state-of-the-art alignment results for a wide range of Western music.
 Using suitable feature representations and cost measures, the toolbox's core technology is based on *dynamic time warping* (DTW), which brings the feature sequences into temporal correspondence.
 To account for efficiency, robustness and, accuracy, our toolbox integrates and combines techniques such as multiscale DTW (MsDTW) [@MuellerMK06_EfficientMultiscaleApproach_ISMIR; @SalvadorC04_fastDTW], 
 memory-restricted MsDTW (MrMsDTW) [@PraetzlichDM16_MsDTW_ICASSP], 
 and high-resolution music synchronization [@EwertMG09_HighResAudioSync_ICASSP].
-While realizing a complete system with presets that allow users to reproduce research results from the literature, our toolbox also provides well-document functions for all required basic building blocks for feature extraction and alignment.
+While realizing a complete system with presets that allow users to reproduce research results from the literature, our toolbox also provides well-documented functions for all basic building blocks required for feature extraction and alignment.
 Furthermore, the toolbox contains example code for visualizing, sonifying, and evaluating synchronization results, thus deepening the understanding of the techniques and data.
 
 # Statement of Need
-The task of finding an alignment between two feature sequences has received large research interest in the past, in the context of MIR and beyond. In the music domain, alignment techniques are central for applications such as score following, content-based retrieval, automatic accompaniment, or performance analysis [@Arzt16_MusicTracking_PhD; @Mueller15_FMP_SPRINGER].
-Beyond these classical applications, alignment techniques have gained in importance in view of recent data-driven machine learning techniques. In particular, music synchronization has shown its potential for facilitating data annotation, data augmentation, and model evaluation. 
[…]
