### 2021-09-13 b87bda6
documentation update (include pitch, filterbank and utils)
