+++ docs/source/feature/pitch.rst
@@ -1,0 +1,6 @@
+Pitch Features (synctoolbox.feature.pitch)
+==========================================
+
+.. automodule:: synctoolbox.feature.pitch
+    :members:
+    :undoc-members:
