+++ docs/source/feature/filterbank.rst
@@ -1,0 +1,6 @@
+Filterbank Generation (synctoolbox.feature.filterbank)
+======================================================
+
+.. automodule:: synctoolbox.feature.filterbank
+    :members:
+    :undoc-members:
