+++ docs/source/feature/utils.rst
@@ -1,0 +1,6 @@
+Utils (synctoolbox.feature.utils)
+=================================
+
+.. automodule:: synctoolbox.feature.utils
+    :members:
+    :undoc-members:
