### 2021-09-17 abb715b
temporal decay added to the spectral flux.
