+++ synctoolbox/feature/novelty.py
@@ -7,7 +7,8 @@ def spectral_flux(f_audio: np.ndarray,
                   Fs: int = 22050,
                   feature_rate: int = 50,
                   gamma: float = 10,
-                  M_sec: float = 0.1) -> np.ndarray:
+                  M_sec: float = 0.1,
+                  filter_coeff: np.ndarray = np.sqrt(1 / np.arange(1, 11))) -> np.ndarray:
     """Generates the spectral-based novelty curve given an audio array.
 
     This function is based on the FMP notebook on "Spectral-Based Novelty":
@@ -30,10 +31,13 @@ def spectral_flux(f_audio: np.ndarray,
     M_sec: float
         Determines size (2M+1) in samples of centric window  used for local average
 
+    filter_coeff: np.ndarray
+        Sequence of decay coefficients applied on normalized chroma onsets.
+
     Returns
     -------
-    nov_norm : np.ndarray [shape=(N, )]
-        Enhanced novelty curve with the subtraction of a local average
+    sf : np.ndarray [shape=(N, )]
+        Enhanced novelty curve with the subtraction of a local averagenad a temporal decay
     """
 
     window_size = int(Fs / feature_rate * 2)
@@ -63,4 +67,14 @@ def spectral_flux(f_audio: np.ndarray,
     nov_norm[nov_norm < 0] = 0
     nov_norm = nov_norm / max(nov_norm)
 
-    return nov_norm
+    # Add a temporal decay to the novelty curve.
+    v_shift = np.array(nov_norm, copy=True)
+    v_help = np.zeros((nov_norm.shape[0], 10))
+
+    for n in range(len(filter_coeff)):
+        v_help[:, n] = filter_coeff[n] * v_shift
+        v_shift = np.roll(v_shift, 1)
+        v_shift[0] = 0
+
+    sf = np.max(v_help, axis=1)
+    return sf
