### 2021-10-15 fdd45b6
Improved description in the example notebooks; Added a warning in compute_optimal_chroma_shift if this is called on very large inputs; Put evaluation measures for warped (measure) positions into a helper function; Fixed a bug in warping the score note durations in the sync_audio_score_full notebook.
