### 2021-10-19 849db72
added conda environment file
