+++ environment.yml
@@ -1,0 +1,13 @@
+name: synctoolbox
+
+channels:
+  - defaults
+  - conda-forge
+
+dependencies:
+  - python=3.7.*
+  - ipython=7.8.*
+  - jupyter=1.0.*  
+
+  - pip:
+    - synctoolbox==1.1.*
