### 2021-11-01 60c5acb
change logo in the docs
