### 2021-11-02 4229e3d
Several improvements to visualizations of feature extraction and MrMsDTW
