### 2021-11-02 443d0d1
Implemented logic for percussive events in symbolic files
