### 2021-11-02 1fa99d3
Version bump to 1.2.0
