### 2021-11-24 f2cd093
log compression added for the visualization of pitch & pitch onset features.
