+++ synctoolbox/feature/csv_tools.py
@@ -1,3 +1,8 @@
+import os.path
+import subprocess
+
+import libfmp.c1
+import music21
 import numpy as np
 import pandas as pd
 from synctoolbox.feature.pitch import __visualize_pitch
@@ -326,3 +331,115 @@ def __get_audio_duration_from_df(df: pd.DataFrame) -> float:
     duration = df['end'].max()
     return duration
 
+
+def music_xml_to_csv_musical_time(xml, csv_filepath: str):
+    """Convert a music xml file to a list of note events, with starts and durations as fractions of measures, and stores
+    it as a csv file in libfmp format.
+
+    Args:
+        xml (str or music21.stream.Score): Either a path to a music xml file or a music21.stream.Score
+        csv_filepath : str
+            Filepath to the .csv file.
+    """
+
+    if isinstance(xml, str):
+        xml_data = music21.converter.parse(xml)
+    elif isinstance(xml, music21.stream.Score):
+        xml_data = xml
+    else:
+        raise RuntimeError('xml must be a path to an xml file or music21.stream.Score')
+
+    # Convert tied notes to single note events
+    xml_data = xml_data.stripTies()
+    # Expand any repeat signs in the score. Unfortunately, for older versions of music21, this method throws errors if there are no repeats in the score, so we do some try-and-catch: https://github.com/cuthbertLab/music21/issues/355
+    try:
+        xml_data_expanded = xml_data.expandRepeats()
+    except music21.repeat.ExpanderException:
+        xml_data_expanded = xml_data
+    xml_data = xml_data_expanded
+
+    score = []
+    # First, get starts and ends of notes in terms of quarters (similar to 'xml_to_list' in libfmp)
+    for part in xml_data.parts:
+        instrument = part.getInstrument().partName  # Other interesting properties may be instrumentName and midiProgram, but often, these are not set correctly
+        for note in part.flat.notes:
+            start = note.offset
+            end = note.offset + note.quarterLength
+            if note.isChord:
+                for chord_note in note.pitches:
+                    pitch = chord_note.ps
+                    volume = note.volume.realized
+                    score.append([start, end, pitch, volume, instrument])
+            else:
+                pitch = note.pitch.ps
+                volume = note.volume.realized
+                score.append([start, end, pitch, volume, instrument])
+
+    # To get the positions of measures in the score in terms of quarters, make sure that...
+    measures_in_parts = [[measure for measure in part.getElementsByClass(music21.stream.Measure)] for part in xml_data.parts]
+    measure_positions_in_quarters = [[measure.offset for measure in part] for part in measures_in_parts]
+    # ...all parts have the same number of measures and...
+    assert all(len(measure_positions_in_quarters[0]) == len(ms) for ms in measure_positions_in_quarters), "Malformed musicxml: Not all parts have the same number of measures. Sometimes this can be fixed by im- and then exporting the file in musescore."
+    # ...that the positions of measures (in terms of quarters) are the same everywhere.
+    assert all(measure_positions_in_quarters[0][m] == measure_positions_in_quarters[i][m] for i in range(len(xml_data.parts)) for m in range(len(measure_positions_in_quarters[0]))), "Malformed musicxml: Not all parts have the same measure positions. Sometimes this can be fixed by im- and then exporting the file in musescore."  # This happens for example for second piece in SWD.
+
+    measure_positions_in_quarters = measure_positions_in_quarters[0]
+    if measures_in_parts[0][0].duration.quarterLength < measures_in_parts[0][0].barDuration.quarterLength:
+        has_pickup_measure = True  # Start counting at 0
+    else:
+        has_pickup_measure = False  # Start counting at 1
[…]
